# encoding/asn1: improve memory efficiency of ObjectIdentifier.String